requirements.txt
fastapi
openai
python-dotenv
langchain
bs4
tiktoken